backend/requirements.txt
fastapi==0.115.5
openai==1.55.3
python-dotenv==1.0.1
python-multipart==0.0.20
python-pptx==1.0.2
flake8==7.1.1
uvicorn==0.32.1
